checkpoint split up embed.py
