+++ bokeh/embed/__init__.py
@@ -1,0 +1,52 @@
+#-----------------------------------------------------------------------------
+# Copyright (c) 2012 - 2017, Anaconda, Inc. All rights reserved.
+#
+# Powered by the Bokeh Development Team.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+'''
+
+'''
+
+#-----------------------------------------------------------------------------
+# Boilerplate
+#----------------------------------------------------------------------------
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+import logging
+log = logging.getLogger(__name__)
+
+from bokeh.util.api import public, internal ; public, internal
+
+#-----------------------------------------------------------------------------
+# Imports
+#-----------------------------------------------------------------------------
+
+# Standard library imports
+
+# External imports
+
+# Bokeh imports
+
+#-----------------------------------------------------------------------------
+# Public API
+#-----------------------------------------------------------------------------
+
+from .notebook import notebook_content
+
+from .server import server_document
+from .server import server_session
+
+from .standalone import autoload_static
+from .standalone import components
+from .standalone import file_html
+
+__all__ = (
+    'autoload_static',
+    'components',
+    'file_html',
+    'notebook_content',
+    'server_document',
+    'server_session',
+)
