+++ bokeh/tests/test_objects.py
@@ -5,8 +5,6 @@ from six.moves import xrange
 import copy
 from bokeh.core.properties import List, String, Instance, Dict, Any, Int
 from bokeh.model import Model
-from bokeh.embed import _ModelInDocument
-from bokeh.document import Document
 from bokeh.core.property.containers import PropertyValueList, PropertyValueDict
 from bokeh.util.future import with_metaclass
 
@@ -303,75 +301,6 @@ class TestModel(unittest.TestCase):
         # 'child' is a default, but it gets included as a
         # non-default because it's unstable.
         self.assertTrue('child' in obj1.properties_with_values(include_defaults=False))
-
-class SomeModelInTestObjects(Model):
-    child = Instance(Model)
-
-class TestModelInDocument(unittest.TestCase):
-    def test_single_model(self):
-        p = Model()
-        self.assertIs(p.document, None)
-        with _ModelInDocument([p]):
-            self.assertIsNot(p.document, None)
-        self.assertIs(p.document, None)
-
-    def test_list_of_model(self):
-        p1 = Model()
-        p2 = Model()
-        self.assertIs(p1.document, None)
-        self.assertIs(p2.document, None)
-        with _ModelInDocument([p1, p2]):
-            self.assertIsNot(p1.document, None)
-            self.assertIsNot(p2.document, None)
-        self.assertIs(p1.document, None)
-        self.assertIs(p2.document, None)
-
-    def test_uses_precedent(self):
-        # it's deliberate that the doc is on p2, so _ModelInDocument
-        # has to be smart about looking for a doc anywhere in the list
-        # before it starts inventing new documents
-        doc = Document()
-        p1 = Model()
-        p2 = Model()
-        doc.add_root(p2)
-        self.assertIs(p1.document, None)
-        self.assertIsNot(p2.document, None)
-        with _ModelInDocument([p1, p2]):
-            self.assertIsNot(p1.document, None)
-            self.assertIsNot(p2.document, None)
-            self.assertIs(p1.document, doc)
-            self.assertIs(p2.document, doc)
-        self.assertIs(p1.document, None)
-        self.assertIsNot(p2.document, None)
-
-    def test_uses_doc_precedent(self):
-        doc = Document()
-        p1 = Model()
-        p2 = Model()
-        self.assertIs(p1.document, None)
-        self.assertIs(p2.document, None)
-        with _ModelInDocument([p1, p2, doc]):
-            self.assertIsNot(p1.document, None)
-            self.assertIsNot(p2.document, None)
-            self.assertIs(p1.document, doc)
-            self.assertIs(p2.document, doc)
-        self.assertIs(p1.document, None)
-        self.assertIs(p2.document, None)
-
-    def test_with_doc_in_child_raises_error(self):
-        doc = Document()
-        p1 = Model()
-        p2 = SomeModelInTestObjects(child=Model())
-        doc.add_root(p2.child)
-        self.assertIs(p1.document, None)
-        self.assertIs(p2.document, None)
-        self.assertIs(p2.child.document, doc)
-        with self.assertRaisesRegexp(RuntimeError, p2._id):
-            with _ModelInDocument([p1, p2]):
-                self.assertIsNot(p1.document, None)
-                self.assertIsNot(p2.document, None)
-                self.assertIs(p1.document, doc)
-                self.assertIs(p2.document, doc)
 
 class TestContainerMutation(unittest.TestCase):
